Move Resteasy providers into a common subproject
These are registered via `META-INF/services/javax.ws.rs.ext.Providers`
which is supported in both spring boot resteasy and quarkus resteasy :-)
