+++ settings.gradle
@@ -17,6 +17,7 @@ include ':clients:subscription-client'
 include ':clients:swatch-internal-subscription-client'
 include ':clients:user-client'
 include ':clients:rh-partner-gateway-client'
+include ':swatch-common-resteasy'
 include ':swatch-core'
 include ':swatch-core-test'
 include ':swatch-system-conduit'
