+++ swatch-common-resteasy/build.gradle
@@ -1,0 +1,11 @@
+plugins {
+    id 'swatch.java-conventions'
+}
+
+dependencies {
+    compileOnly platform(libraries["quarkus-bom"])
+    compileOnly 'org.jboss.spec.javax.ws.rs:jboss-jaxrs-api_2.1_spec'
+    compileOnly 'org.slf4j:slf4j-api'
+}
+
+description = 'SWATCH common library for RESTEasy configuration & utilities'
