+++ swatch-common-resteasy/src/main/java/com/redhat/swatch/common/resteasy/EnumParamConverterProvider.java
@@ -18,29 +18,21 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.subscriptions.resteasy;
+package com.redhat.swatch.common.resteasy;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
 import java.util.Arrays;
 import java.util.Map;
 import java.util.Objects;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import javax.ws.rs.ext.ParamConverter;
 import javax.ws.rs.ext.ParamConverterProvider;
-import javax.ws.rs.ext.Provider;
 import org.slf4j.MDC;
-import org.springframework.stereotype.Component;
 
-/** ParamConverterProvider to enable use of enums in query providers. */
-@Component
-@Provider
+/** ParamConverterProvider to enable use of enums in query params. */
 public class EnumParamConverterProvider implements ParamConverterProvider {
-
   public static final String INVALID_VALUE_EXCEPTION_MSG = "%s is not a valid value for %s";
-  private static final String PACKAGE_CONTAINING_SPECIAL_ENUMS =
-      "org.candlepin.subscriptions.utilization.api.model";
 
   @SuppressWarnings("linelength")
   @Override
@@ -50,13 +42,11 @@ public class EnumParamConverterProvider implements ParamConverterProvider {
       return null;
     }
 
-    return rawType.getPackage().getName().startsWith(PACKAGE_CONTAINING_SPECIAL_ENUMS)
-        ? new EnumParamConverter<>(rawType, false)
-        : new EnumParamConverter<>(rawType);
+    return new EnumParamConverter<>(rawType);
   }
 
   /**
-   * ParamConverter that uses an ObjectMapper to convert enums to/from String.
+   * ParamConverter that case-insensitively matches enums
    *
    * @param <T> the enum type
    */
@@ -64,26 +54,13 @@ public class EnumParamConverterProvider implements ParamConverterProvider {
 
     private final Class<T> className;
     private final Map<String, T> stringToEnumMap;
-    private final boolean isCaseSensitive;
 
     public EnumParamConverter(Class<T> className) {
-      this(className, true);
-    }
-
-    public EnumParamConverter(Class<T> className, boolean isCaseSensitive) {
       this.className = className;
-      this.isCaseSensitive = isCaseSensitive;
-
-      Function<T, String> applyToStringFunction =
-          isCaseSensitive ? T::toString : t -> t.toString().toLowerCase();
 
       stringToEnumMap =
           Arrays.stream(className.getEnumConstants())
-              .collect(Collectors.toMap(applyToStringFunction, value -> value));
-    }
-
-    public Class<T> getClassName() {
-      return className;
+              .collect(Collectors.toMap(value -> value.toString().toLowerCase(), value -> value));
     }
 
     public T fromString(String value) {
@@ -91,7 +68,7 @@ public class EnumParamConverterProvider implements ParamConverterProvider {
         return null;
       }
 
-      String lookupValue = isCaseSensitive ? value : value.toLowerCase();
+      String lookupValue = value.toLowerCase();
 
       T result = stringToEnumMap.get(lookupValue);
       if (result != null) {
