+++ swatch-common-resteasy/src/main/resources/META-INF/services/javax.ws.rs.ext.Providers
@@ -1,0 +1,2 @@
+com.redhat.swatch.common.resteasy.EnumParamConverterProvider
+com.redhat.swatch.common.resteasy.OffsetDateTimeParamConverterProvider
