+++ swatch-contracts/build.gradle
@@ -25,6 +25,7 @@ dependencies {
     implementation 'io.quarkus:quarkus-smallrye-health'
     implementation 'io.quarkus:quarkus-smallrye-openapi'
     implementation project(':clients:swatch-internal-subscription-client')
+    implementation project(':swatch-common-resteasy')
     implementation libraries["clowder-quarkus-config-source"]
     implementation libraries["quarkus-logging-splunk"]
     implementation libraries["splunk-library-javalogging"]
